# docs/openapi-old-multiframework.yaml
openapi: 3.0.3
info:
  title: MCP Python API
  description: |
    Comprehensive Multi-Framework AI Platform API

    This API provides unified access to multiple AI frameworks including:
    - Haystack for retrieval-augmented generation
    - LlamaIndex for document indexing and querying
    - LangChain for chains and agents
    - LangGraph for stateful workflows
    - OpenAI for direct API access
    - Transformers for local model inference
    - Anthropic Claude for advanced reasoning
    - Cohere for text generation and embeddings
    - Custom framework adapters

    Features:
    - Multi-framework query processing
    - Advanced search and filtering
    - Real-time analytics and monitoring
    - Comprehensive agent management
    - Document processing and RAG
    - Stateful workflow execution
    - Custom framework integration

  version: 1.0.0
  contact:
    name: MCP Python API Support
    email: support@mcp-python.com
  license:
    name: Proprietary License
    url: https://github.com/wavewand/wand/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.mcp-python.com
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /langchain/agents:
    get:
      summary: List LangChain agents
      tags: [LangChain]
      responses:
        '200':
          description: List of available agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentInfo'
    post:
      summary: Create LangChain agent
      tags: [LangChain]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
...
  /agents/{agentId}/execute:
    post:
      summary: Execute agent task
      tags: [Agents]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecuteRequest'
      responses:
        '200':
          description: Agent execution results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentExecuteResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AgentInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        framework:
          type: string
        status:
          type: string
        tools:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required:
        - name
        - type
        - tools
      properties:
        name:
          type: string
        type:
          type: string
        tools:
          type: array
          items:
            type: string
        config:
          type: object
    AgentExecuteRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
        parameters:
          type: object
    AgentExecuteResponse:
      type: object
      properties:
        success:
          type: boolean
        result:
          type: string
        steps:
          type: array
          items:
            type: object
        execution_time_ms:
          type: number

    # Workflow Management
...